Lets make it simpler and remove router
