+++ package.json
@@ -9,7 +9,6 @@
     "prop-types": "^15.6.1",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
-    "react-router-dom": "^4.3.1",
     "react-scripts": "1.1.1",
     "styled-components": "^3.2.1"
   },
