+++ src/App.js
@@ -1,19 +1,89 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React, { Component } from 'react';
+import styled from 'styled-components';
+import isEmpty from 'lodash.isempty';
 
-const App = ({ children }) => (
-  <section>{children}</section>
-);
+// components:
+import Marker from './components/Marker';
+import GoogleMap from './components/GoogleMap';
 
-App.propTypes = {
-  children: PropTypes.oneOfType([
-    PropTypes.node,
-    PropTypes.arrayOf(PropTypes.node),
-  ]),
+import LOS_ANGELES_CENTER from './const/la_center';
+
+const Wrapper = styled.section`
+  width: 100vw;
+  height: 100vh;
+`;
+
+// Return map bounds based on list of places
+const getMapBounds = (map, maps, places) => {
+  const bounds = new maps.LatLngBounds();
+
+  places.forEach((place) => {
+    bounds.extend(new maps.LatLng(
+      place.geometry.location.lat,
+      place.geometry.location.lng,
+    ));
+  });
+  return bounds;
 };
 
-App.defaultProps = {
-  children: null,
+// Re-center map when resizing the window
+const bindResizeListener = (map, maps, bounds) => {
+  maps.event.addDomListenerOnce(map, 'idle', () => {
+    maps.event.addDomListener(window, 'resize', () => {
+      map.fitBounds(bounds);
+    });
+  });
 };
+
+// Fit map to its bounds after the api is loaded
+const apiIsLoaded = (map, maps, places) => {
+  // Get bounds by our places
+  const bounds = getMapBounds(map, maps, places);
+  // Fit map to bounds
+  map.fitBounds(bounds);
+  // Bind the resize listener
+  bindResizeListener(map, maps, bounds);
+};
+
+class App extends Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      places: [],
+    };
+  }
+
+  componentDidMount() {
+    fetch('/places.json')
+      .then(response => response.json())
+      .then(data => this.setState({ places: data.results }));
+  }
+
+  render() {
+    const { places } = this.state;
+    return (
+      <Wrapper>
+        {!isEmpty(places) && (
+          <GoogleMap
+            defaultZoom={10}
+            defaultCenter={LOS_ANGELES_CENTER}
+            yesIWantToUseGoogleMapApiInternals
+            onGoogleApiLoaded={({ map, maps }) => apiIsLoaded(map, maps, places)}
+          >
+            {places.map(place => (
+              <Marker
+                key={place.id}
+                text={place.name}
+                lat={place.geometry.location.lat}
+                lng={place.geometry.location.lng}
+              />
+            ))}
+          </GoogleMap>
+        )}
+      </Wrapper>
+    );
+  }
+}
 
 export default App;
