--- src/modules/Main.js
@@ -1,89 +1,0 @@
-import React, { Component } from 'react';
-import styled from 'styled-components';
-import isEmpty from 'lodash.isempty';
-
-// components:
-import Marker from '../components/Marker';
-import GoogleMap from '../components/GoogleMap';
-
-import LOS_ANGELES_CENTER from '../const/la_center';
-
-const Wrapper = styled.section`
-  width: 100vw;
-  height: 100vh;
-`;
-
-// Return map bounds based on list of places
-const getMapBounds = (map, maps, places) => {
-  const bounds = new maps.LatLngBounds();
-
-  places.forEach((place) => {
-    bounds.extend(new maps.LatLng(
-      place.geometry.location.lat,
-      place.geometry.location.lng,
-    ));
-  });
-  return bounds;
-};
-
-// Re-center map when resizing the window
-const bindResizeListener = (map, maps, bounds) => {
-  maps.event.addDomListenerOnce(map, 'idle', () => {
-    maps.event.addDomListener(window, 'resize', () => {
-      map.fitBounds(bounds);
-    });
-  });
-};
-
-// Fit map to its bounds after the api is loaded
-const apiIsLoaded = (map, maps, places) => {
-  // Get bounds by our places
-  const bounds = getMapBounds(map, maps, places);
-  // Fit map to bounds
-  map.fitBounds(bounds);
-  // Bind the resize listener
-  bindResizeListener(map, maps, bounds);
-};
-
-class Main extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      places: [],
-    };
-  }
-
-  componentDidMount() {
-    fetch('/places.json')
-      .then(response => response.json())
-      .then(data => this.setState({ places: data.results }));
-  }
-
-  render() {
-    const { places } = this.state;
-    return (
-      <Wrapper>
-        {!isEmpty(places) && (
-          <GoogleMap
-            defaultZoom={10}
-            defaultCenter={LOS_ANGELES_CENTER}
-            yesIWantToUseGoogleMapApiInternals
-            onGoogleApiLoaded={({ map, maps }) => apiIsLoaded(map, maps, places)}
-          >
-            {places.map(place => (
-              <Marker
-                key={place.id}
-                text={place.name}
-                lat={place.geometry.location.lat}
-                lng={place.geometry.location.lng}
-              />
-            ))}
-          </GoogleMap>
-        )}
-      </Wrapper>
-    );
-  }
-}
-
-export default Main;
